+++ src/components/user-settings/PrivacySetting.js
@@ -3,10 +3,10 @@ import { observer } from 'mobx-react';
 import { View, Text, StyleSheet, AlertIOS, Switch } from 'react-native';
 import { Icon } from 'react-native-elements';
 
-import { container as containerStyles } from '../styles';
-import { colors, divider } from '../styles/theme';
-import UserStore from '../stores/user';
-import FormHeader from '../custom/FormHeader';
+import { container as containerStyles } from '../../styles';
+import { colors, divider } from '../../styles/theme';
+import UserStore from '../../stores/user';
+import FormHeader from '../../custom/FormHeader';
 
 @observer
 export class PrivacySetting extends React.Component {
